I updated the Fontend
I added in the results page and added in a footer that contains our names.
I also fixed the styling of the website so tat it could work with the logic of hidding and showing sections of the game.
Currenly solving an issue with the playing again button for the results page.