Updated some of the classes and added html folder
I updated some of the classes and fix some of them so that the program would be able to compile. i also removed the Test java files since for the assignemt its not yet needed. i removed  the MessageHandler file since its not one of our classes from the class diagram.